+++ lib/src/ast/css/keyframe_block.dart
@@ -2,24 +2,13 @@
 // MIT-style license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
-import 'package:source_span/source_span.dart';
-
-import '../../visitor/interface/css.dart';
 import 'node.dart';
 import 'value.dart';
 
 /// A block within a `@keyframes` rule.
 ///
 /// For example, `10% {opacity: 0.5}`.
-class CssKeyframeBlock extends CssParentNode {
+abstract class CssKeyframeBlock extends CssParentNode {
   /// The selector for this block.
-  final CssValue<List<String>> selector;
-
-  final FileSpan span;
-
-  CssKeyframeBlock(this.selector, this.span);
-
-  T accept<T>(CssVisitor<T> visitor) => visitor.visitKeyframeBlock(this);
-
-  CssKeyframeBlock copyWithoutChildren() => CssKeyframeBlock(selector, span);
+  CssValue<List<String>> get selector;
 }